adicionado os valores iniciais

+++ status.h
@@ -8,7 +8,7 @@ extern "C"{
 typedef struct {
   uint16_t c_angulo_final;
   uint16_t c_angulo_inicial;
-  uint32_t c_angulo_encoder;
+  uint16_t c_angulo_encoder;
 
   uint16_t c_tempo_insp_cont;
   uint16_t c_tempo_insp_IHM;
